Bug 958375 - 8/9 - Make remaining Moz2D enums typed - r=Bas
Specifically:
  r=Bas for manual changes
  f=Bas for automatic changes
See attachments on the bug for the specific breakdown.

+++ gfx/thebes/gfxPlatform.h
@@ -54,7 +54,7 @@ class DrawEventRecorder;
 inline uint32_t
 BackendTypeBit(BackendType b)
 {
-  return 1 << int(b);
+  return 1 << uint8_t(b);
 }
 }
 }
